Read source file relative from base directory
Given a path from a script containing a sourcemap url of /index.js.map and a directory of /src/project/a/ where the index.js.map file lives within, the location would resolve to /index.js on the file system. This uses /src/project/a/ as an "absolute" base so that we resolve to /src/project/a/index.js.map.